% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl.R
\name{build_ardl_design}
\alias{build_ardl_design}
\title{Build ARDL Design Matrix}
\usage{
build_ardl_design(y, X, fourier, p, q)
}
\arguments{
\item{y}{Dependent variable}

\item{X}{Independent variables matrix}

\item{fourier}{Fourier terms}

\item{p}{Lag order for y}

\item{q}{Lag order for X}
}
\value{
List with y and X design matrices
}
\description{
Constructs the design matrix for ARDL estimation including
lagged dependent and independent variables.
}
\keyword{internal}
