% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtnardl.R
\name{build_mtnardl_regressors}
\alias{build_mtnardl_regressors}
\title{Build MTNARDL Regressor Matrix}
\usage{
build_mtnardl_regressors(data, x_names, decompose, decomposed, regime_names)
}
\description{
Build MTNARDL Regressor Matrix
}
\keyword{internal}
