% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnardl.R
\name{build_nardl_regressors}
\alias{build_nardl_regressors}
\title{Build NARDL Regressor Matrix}
\usage{
build_nardl_regressors(data, x_names, decompose, decomposed_data)
}
\arguments{
\item{data}{Original data}

\item{x_names}{All independent variable names}

\item{decompose}{Variables that are decomposed}

\item{decomposed_data}{Output from decompose_variables}
}
\value{
Matrix of regressors
}
\description{
Build NARDL Regressor Matrix
}
\keyword{internal}
