% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnardl.R
\name{compute_asymmetric_multipliers}
\alias{compute_asymmetric_multipliers}
\title{Compute Asymmetric Multipliers}
\usage{
compute_asymmetric_multipliers(nardl_result, decompose, x_names)
}
\arguments{
\item{nardl_result}{NARDL estimation results}

\item{decompose}{Decomposed variable names}

\item{x_names}{Original variable names}
}
\value{
List with long-run and short-run asymmetric multipliers
}
\description{
Compute Asymmetric Multipliers
}
