% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl.R
\name{compute_diagnostics}
\alias{compute_diagnostics}
\title{Compute Model Diagnostics}
\usage{
compute_diagnostics(qardl_results)
}
\arguments{
\item{qardl_results}{List of QARDL estimation results}
}
\value{
List of diagnostics for each quantile
}
\description{
Computes various diagnostic statistics for the QARDL models.
}
