% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl.R
\name{compute_multipliers}
\alias{compute_multipliers}
\title{Compute Long-run and Short-run Multipliers}
\usage{
compute_multipliers(qardl_results, x_names, tau)
}
\arguments{
\item{qardl_results}{List of QARDL results for each quantile}

\item{x_names}{Names of independent variables}

\item{tau}{Vector of quantiles}
}
\value{
List with long-run and short-run multiplier matrices
}
\description{
Calculates the long-run and short-run multipliers from QARDL estimates.
}
