% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtnardl.R
\name{compute_regime_multipliers}
\alias{compute_regime_multipliers}
\title{Compute Regime-Specific Multipliers}
\usage{
compute_regime_multipliers(model_result, decompose, regime_names)
}
\description{
Compute Regime-Specific Multipliers
}
\keyword{internal}
