% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnardl.R
\name{decompose_variables}
\alias{decompose_variables}
\title{Decompose Variables into Positive and Negative Changes}
\usage{
decompose_variables(data, variables)
}
\arguments{
\item{data}{Data frame}

\item{variables}{Variables to decompose}
}
\value{
List with positive and negative components
}
\description{
Decomposes time series into cumulative positive and negative partial sums.
}
