% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnardl.R
\name{estimate_nardl}
\alias{estimate_nardl}
\title{Estimate NARDL Model}
\usage{
estimate_nardl(y, X, fourier, p, q, case = 3)
}
\arguments{
\item{y}{Dependent variable}

\item{X}{Regressor matrix (with decomposed variables)}

\item{fourier}{Fourier terms}

\item{p}{Lag for y}

\item{q}{Lag for X}

\item{case}{Model case}
}
\value{
List with estimation results
}
\description{
Estimate NARDL Model
}
\keyword{internal}
