% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl.R
\name{estimate_qardl}
\alias{estimate_qardl}
\title{Estimate Quantile ARDL Model}
\usage{
estimate_qardl(y, X, fourier, p, q, tau, case = 3)
}
\arguments{
\item{y}{Dependent variable}

\item{X}{Independent variables}

\item{fourier}{Fourier terms}

\item{p}{Lag for dependent variable}

\item{q}{Lag for independent variables}

\item{tau}{Quantile (0 < tau < 1)}

\item{case}{Model case (1-5)}
}
\value{
List with estimation results
}
\description{
Estimates the Quantile ARDL model for a given quantile tau.
}
