% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funitroot.R
\name{get_fadf_critical_values}
\alias{get_fadf_critical_values}
\title{Get Fourier ADF Critical Values}
\usage{
get_fadf_critical_values(n, model, k)
}
\description{
Get Fourier ADF Critical Values
}
\keyword{internal}
