% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funitroot.R
\name{get_fkpss_critical_values}
\alias{get_fkpss_critical_values}
\title{Get Fourier KPSS Critical Values}
\usage{
get_fkpss_critical_values(model, k, n = 100)
}
\description{
Get Fourier KPSS Critical Values
}
\keyword{internal}
