% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_3d_surface}
\alias{plot_3d_surface}
\title{Plot 3D Surface of Coefficients}
\usage{
plot_3d_surface(obj, variable = NULL)
}
\arguments{
\item{obj}{FQARDL object}

\item{variable}{Variable to plot}
}
\value{
plotly object
}
\description{
Plot 3D Surface of Coefficients
}
\keyword{internal}
