% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_nardl.R
\name{plot_asymmetry}
\alias{plot_asymmetry}
\title{Plot Asymmetry Bar Chart}
\usage{
plot_asymmetry(obj, variable = NULL)
}
\arguments{
\item{obj}{FNARDL object}

\item{variable}{Variable to plot}
}
\description{
Plot Asymmetry Bar Chart
}
\keyword{internal}
