% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_coefficients}
\alias{plot_coefficients}
\title{Plot Coefficients Across Quantiles}
\usage{
plot_coefficients(obj, variable = NULL)
}
\arguments{
\item{obj}{FQARDL object}

\item{variable}{Variable name (NULL for all)}
}
\value{
ggplot object
}
\description{
Plot Coefficients Across Quantiles
}
\keyword{internal}
