% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_nardl.R
\name{plot_cumulative_multipliers}
\alias{plot_cumulative_multipliers}
\title{Plot Cumulative Multipliers}
\usage{
plot_cumulative_multipliers(obj, variable, horizon = 20)
}
\arguments{
\item{obj}{FNARDL object}

\item{variable}{Variable to plot}

\item{horizon}{Number of periods}
}
\description{
Plot Cumulative Multipliers
}
\value{No return value, called for side effects (generates output or plots)}
