% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap of Coefficients}
\usage{
plot_heatmap(obj)
}
\arguments{
\item{obj}{FQARDL object}
}
\value{
ggplot object
}
\description{
Plot Heatmap of Coefficients
}
\keyword{internal}
