% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_multipliers}
\alias{plot_multipliers}
\title{Plot Long-run and Short-run Multipliers}
\usage{
plot_multipliers(obj)
}
\arguments{
\item{obj}{FQARDL object}
}
\value{
ggplot object
}
\description{
Plot Long-run and Short-run Multipliers
}
\keyword{internal}
