% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_persistence}
\alias{plot_persistence}
\title{Plot Persistence Profile}
\usage{
plot_persistence(obj, horizons = 20)
}
\arguments{
\item{obj}{FQARDL object}

\item{horizons}{Number of periods for persistence profile}
}
\value{
ggplot object
}
\description{
Plots the persistence profile showing the adjustment path
to long-run equilibrium after a shock.
}
