% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_nardl.R
\name{plot_pos_neg_comparison}
\alias{plot_pos_neg_comparison}
\title{Plot Positive vs Negative Comparison}
\usage{
plot_pos_neg_comparison(obj)
}
\arguments{
\item{obj}{FNARDL object}
}
\description{
Plot Positive vs Negative Comparison
}
\keyword{internal}
