% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Residual Diagnostics}
\usage{
plot_residuals(obj)
}
\arguments{
\item{obj}{FQARDL object}
}
\value{
ggplot object
}
\description{
Plot Residual Diagnostics
}
\keyword{internal}
