% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_test.R
\name{quantile_wald_test}
\alias{quantile_wald_test}
\title{Quantile Wald Test for Coefficient Constancy}
\usage{
quantile_wald_test(qardl_results, coef_name)
}
\arguments{
\item{qardl_results}{List of QARDL results}

\item{coef_name}{Name of coefficient to test}
}
\value{
List with test results
}
\description{
Tests whether coefficients are constant across quantiles.
}
