% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{select_fourier_frequency}
\alias{select_fourier_frequency}
\title{Select Optimal Fourier Frequency}
\usage{
select_fourier_frequency(y, X, max_k = 3, criterion = c("BIC", "AIC", "HQ"))
}
\arguments{
\item{y}{Dependent variable vector}

\item{X}{Matrix of independent variables}

\item{max_k}{Maximum Fourier frequency to test}

\item{criterion}{Information criterion ("AIC", "BIC", "HQ")}
}
\value{
A list containing:
\item{optimal_k}{The optimal Fourier frequency}
\item{ic_values}{Information criterion values for each k}
\item{criterion}{The criterion used}
}
\description{
Selects the optimal Fourier frequency k based on information criteria.
Tests all frequencies from 1 to max_k and selects the one minimizing
the chosen criterion.
}
