% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtnardl.R
\name{select_mtnardl_lags}
\alias{select_mtnardl_lags}
\title{Select Optimal Lags for MTNARDL}
\usage{
select_mtnardl_lags(y, X, max_p, max_q, criterion)
}
\description{
Select Optimal Lags for MTNARDL
}
\keyword{internal}
