% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl.R
\name{select_optimal_lags}
\alias{select_optimal_lags}
\title{Select Optimal Lag Structure}
\usage{
select_optimal_lags(
  y,
  X,
  fourier,
  max_p,
  max_q,
  criterion = c("BIC", "AIC", "HQ")
)
}
\arguments{
\item{y}{Dependent variable}

\item{X}{Matrix of independent variables}

\item{fourier}{Fourier terms matrix}

\item{max_p}{Maximum lag for dependent variable}

\item{max_q}{Maximum lag for independent variables}

\item{criterion}{Information criterion}
}
\value{
List with optimal lags
}
\description{
Selects optimal lag orders for ARDL model using grid search
over all combinations of p and q.
}
