% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnardl.R
\name{test_asymmetry}
\alias{test_asymmetry}
\title{Test for Asymmetry (Wald Test)}
\usage{
test_asymmetry(nardl_result, decompose)
}
\arguments{
\item{nardl_result}{NARDL estimation results}

\item{decompose}{Decomposed variables}
}
\value{
List of Wald test results for each variable
}
\description{
Test for Asymmetry (Wald Test)
}
