% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_test.R
\name{bootstrap_bounds_test}
\alias{bootstrap_bounds_test}
\title{Bootstrap Bounds Test}
\usage{
bootstrap_bounds_test(
  y,
  X,
  fourier,
  p,
  q,
  tau,
  case,
  n_boot = 1000,
  verbose = FALSE
)
}
\arguments{
\item{y}{Dependent variable}

\item{X}{Independent variables}

\item{fourier}{Fourier terms}

\item{p}{Lag for y}

\item{q}{Lag for X}

\item{tau}{Quantiles}

\item{case}{Model case}

\item{n_boot}{Number of bootstrap replications}

\item{verbose}{Logical. Print progress messages (default: FALSE)}
}
\value{
List with bootstrap p-values
}
\description{
Performs bootstrap-based bounds test for cointegration,
following McNown et al. (2018) methodology.
}
