% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funitroot.R
\name{fadf_f_test}
\alias{fadf_f_test}
\title{F-test for Linearity in Fourier ADF}
\usage{
fadf_f_test(y, model, k, p)
}
\arguments{
\item{y}{Time series}

\item{model}{Model specification}

\item{k}{Fourier frequency}

\item{p}{Number of lags}
}
\value{
List with F-statistic and p-value
}
\description{
Tests H0: gamma1 = gamma2 = 0 (no Fourier terms needed)
}
