% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{fourier_adf}
\alias{fourier_adf}
\title{Fourier ADF Unit Root Test}
\usage{
fourier_adf(y, max_k = 3, max_lag = 8, criterion = c("BIC", "AIC"))
}
\arguments{
\item{y}{Time series vector}

\item{max_k}{Maximum Fourier frequency}

\item{max_lag}{Maximum number of lags for ADF}

\item{criterion}{Lag selection criterion}
}
\value{
A list with test results
}
\description{
Performs the Fourier Augmented Dickey-Fuller test for unit roots
with smooth structural breaks.
}
