% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funitroot.R
\name{fourier_kpss_test}
\alias{fourier_kpss_test}
\title{Fourier KPSS Test}
\usage{
fourier_kpss_test(y, model = c("c", "ct"), max_freq = 3, verbose = TRUE)
}
\arguments{
\item{y}{Numeric vector of time series data}

\item{model}{Model specification: "c" (constant), "ct" (constant + trend)}

\item{max_freq}{Maximum Fourier frequency (default: 3)}

\item{verbose}{Logical. Print progress messages (default: TRUE)}
}
\value{
Object of class "fkpss" with test results
}
\description{
Tests for stationarity allowing for smooth structural breaks.
Implements Becker, Enders & Lee (2006) methodology.
}
\references{
Becker, R., Enders, W., & Lee, J. (2006). A stationarity test in the presence
of an unknown number of smooth breaks. Journal of Time Series Analysis, 27(3), 381-409.
}
